Remove version pins on boto3 and botocore.
This is meant to prevent use of moto with the latest botocore since
botocore no longer uses requests for making calls. This change means
that moto can't intercept boto calls and thus tests will call out to the
actual AWS endpoint. However, we are overriding this since we'll use the
deprecated decorators which patch at the socket level and should still
work for testing.

Once moto merges the PR to fix the above issue then we can remove our
fork of moto entirely and just use the latest version.